emater/requirements.txt
Django>=4.0,<5.0
mysqlclient>=2.1,<3.0
pytest>=7.0,<8.0
pytest-django>=4.5,<5.0
coverage>=6.0,<7.0
validate-docbr>=1.10,<2.0
django-bootstrap-v5>=1.0
openpyxl>=3.1.2
djangorestframework>=3.16.0
requests>=2.31.0
